+++ src/metadata/map_storage.rs
@@ -9,8 +9,9 @@
 use crate::{
     chunk_store::{MapChunkStore, UsedSpace},
     error::convert_to_error_message,
+    network::Network,
     node_ops::{NodeDuty, OutgoingMsg},
-    Error, Network, Result,
+    Error, Result,
 };
 use log::info;
 use sn_data_types::{