+++ src/node/member_churn.rs
@@ -9,13 +9,17 @@
 use crate::{
     capacity::{Capacity, ChunkHolderDbs, RateLimit},
     metadata::{adult_reader::AdultReader, Metadata},
+    node::{ElderRole, Role},
     node_ops::NodeDuty,
     section_funds::{reward_wallets::RewardWallets, SectionFunds},
-    transfers::get_replicas::{replica_info, transfer_replicas},
-    transfers::Transfers,
+    transfers::{
+        get_replicas::{replica_info, transfer_replicas},
+        Transfers,
+    },
     Error, Node, Result,
 };
 use crdts::Actor;
+use dashmap::DashMap;
 use itertools::Itertools;
 use log::{debug, info};
 use sn_data_types::{
@@ -33,19 +37,15 @@ impl Node {
     /// If we are an oldie we'll have a transfer instance,
     /// This updates the replica info on it.
     pub async fn update_replicas(&mut self) -> Result<()> {
-        if let Some(ref mut transfers) = self.transfers {
-            let info = replica_info(&self.node_info, &self.network_api).await?;
-            transfers.update_replica_info(info);
-        }
+        let elder = self.role.as_elder_mut()?;
+        let info = replica_info(&self.node_info, &self.network_api).await?;
+        elder.transfers.update_replica_info(info);
         Ok(())
     }
 
     /// Level up a newbie to an oldie on promotion
     pub async fn level_up(&mut self) -> Result<()> {
-        //
-        // do not hande immutable chunks anymore
-        self.chunks = None;
-        self.used_space.reset().await;
+        self.used_space.reset().await; // TODO(drusu): should this be part of adult_state?
 
         //
         // start handling metadata
@@ -53,7 +53,6 @@ impl Node {
         let reader = AdultReader::new(self.network_api.clone());
         let meta_data =
             Metadata::new(&self.node_info.path(), &self.used_space, dbs, reader).await?;
-        self.meta_data = Some(meta_data);
 
         //
         // start handling transfers
@@ -61,15 +60,22 @@ impl Node {
         let rate_limit = RateLimit::new(self.network_api.clone(), Capacity::new(dbs.clone()));
         let user_wallets = BTreeMap::<PublicKey, ActorHistory>::new();
         let replicas = transfer_replicas(&self.node_info, &self.network_api, user_wallets).await?;
-        self.transfers = Some(Transfers::new(replicas, rate_limit));
+        let transfers = Transfers::new(replicas, rate_limit);
 
         //
         // start handling node rewards
-        self.section_funds = Some(SectionFunds::KeepingNodeWallets {
+        let section_funds = SectionFunds::KeepingNodeWallets {
             wallets: RewardWallets::new(BTreeMap::<XorName, (NodeAge, PublicKey)>::new()),
-            payments: Default::default(),
+            payments: DashMap::new(),
+        };
+
+        self.role = Role::Elder(ElderRole {
+            meta_data,
+            transfers,
+            section_funds,
         });
 
+        // TODO(drusu): return a  mutable reference to elder state?
         Ok(())
     }
 
@@ -79,24 +85,14 @@ impl Node {
         node_wallets: BTreeMap<XorName, (NodeAge, PublicKey)>,
         user_wallets: BTreeMap<PublicKey, ActorHistory>,
     ) -> Result<NodeDuty> {
+        let elder = self.role.as_elder_mut()?;
+
         // merge in provided user wallets
-        if let Some(transfers) = &mut self.transfers {
-            transfers.merge(user_wallets)
-        }
+        elder.transfers.merge(user_wallets);
 
         //  merge in provided node reward stages
-        match &mut self.section_funds {
-            Some(SectionFunds::KeepingNodeWallets { wallets, .. })
-            | Some(SectionFunds::Churning { wallets, .. }) => {
-                for (key, (age, wallet)) in &node_wallets {
-                    wallets.set_node_wallet(*key, *age, *wallet);
-                }
-            }
-            None => {
-                return Err(Error::InvalidOperation(
-                    "Invalid section funds stage".to_string(),
-                ))
-            }
+        for (key, (age, wallet)) in &node_wallets {
+            elder.section_funds.set_node_wallet(*key, *wallet, *age)
         }
 
         let node_id = self.network_api.our_name().await;