+++ src/node/mod.rs
@@ -18,12 +18,12 @@ use crate::{
     chunks::Chunks,
     event_mapping::{map_routing_event, LazyError, Mapping, MsgContext},
     metadata::{adult_reader::AdultReader, Metadata},
+    network::Network,
     node_ops::{NodeDuties, NodeDuty},
     section_funds::SectionFunds,
     state_db::store_new_reward_keypair,
-    transfers::get_replicas::transfer_replicas,
-    transfers::Transfers,
-    Config, Error, Network, Result,
+    transfers::{get_replicas::transfer_replicas, Transfers},
+    Config, Error, Result,
 };
 use bls::SecretKey;
 use ed25519_dalek::PublicKey as Ed25519PublicKey;
@@ -65,6 +65,56 @@ impl NodeInfo {
     }
 }
 
+struct AdultRole {
+    // immutable chunks
+    chunks: Chunks,
+}
+
+struct ElderRole {
+    // data operations
+    meta_data: Metadata,
+    // transfers
+    transfers: Transfers,
+    // reward payouts
+    section_funds: SectionFunds,
+}
+
+#[allow(clippy::large_enum_variant)]
+enum Role {
+    Adult(AdultRole),
+    Elder(ElderRole),
+}
+
+impl Role {
+    fn as_adult(&self) -> Result<&AdultRole> {
+        match self {
+            Self::Adult(adult_state) => Ok(adult_state),
+            _ => Err(Error::NotAnAdult),
+        }
+    }
+
+    fn as_adult_mut(&mut self) -> Result<&mut AdultRole> {
+        match self {
+            Self::Adult(adult_state) => Ok(adult_state),
+            _ => Err(Error::NotAnAdult),
+        }
+    }
+
+    fn as_elder(&self) -> Result<&ElderRole> {
+        match self {
+            Self::Elder(elder_state) => Ok(elder_state),
+            _ => Err(Error::NotAnElder),
+        }
+    }
+
+    fn as_elder_mut(&mut self) -> Result<&mut ElderRole> {
+        match self {
+            Self::Elder(elder_state) => Ok(elder_state),
+            _ => Err(Error::NotAnElder),
+        }
+    }
+}
+
 /// Main node struct.
 pub struct Node {
     network_api: Network,
@@ -72,14 +122,7 @@ pub struct Node {
     node_info: NodeInfo,
     used_space: UsedSpace,
     prefix: Prefix,
-    // immutable chunks
-    chunks: Option<Chunks>,
-    // data operations
-    meta_data: Option<Metadata>,
-    // transfers
-    transfers: Option<Transfers>,
-    // reward payouts
-    section_funds: Option<SectionFunds>,
+    role: Role,
 }
 
 impl Node {
@@ -124,21 +167,18 @@ impl Node {
 
         let node = Self {
             prefix: network_api.our_prefix().await,
-            chunks: Some(
-                Chunks::new(
+            role: Role::Adult(AdultRole {
+                chunks: Chunks::new(
                     node_info.node_name,
                     node_info.root_dir.as_path(),
                     used_space.clone(),
                 )
                 .await?,
-            ),
+            }),
             node_info,
             used_space,
             network_api,
             network_events,
-            meta_data: None,
-            transfers: None,
-            section_funds: None,
         };
 
         messaging::send(node.register_wallet().await, &node.network_api).await;